made everything use the functions in the utils file
