+++ client/src/Utils.js
@@ -163,3 +163,70 @@ export class Utils {
       });
     }
   }
+
+  /* ==================== UTILITY FUNCTIONS: Database Schema Management ==================== */
+class DatabaseService {
+    static async query(env, sql, params = []) {
+      try {
+        const statement = env.D1_DB.prepare(sql);
+        params.forEach((param, index) => {
+          statement.bind(param, index + 1);
+        });
+        const { results } = await statement.all();
+        return results;
+      } catch (error) {
+        throw new Error(`Database query failed: ${error.message}`);
+      }
+    }
+  
+    static async execute(env, sql, params = []) {
+      try {
+        const statement = env.D1_DB.prepare(sql);
+        params.forEach((param, index) => {
+          statement.bind(param, index + 1);
+        });
+        const result = await statement.run();
+        return result;
+      } catch (error) {
+        throw new Error(`Database execution failed: ${error.message}`);
+      }
+    }
+  }
+  
+  class BackendService {
+    static async handleRequest(request, handler) {
+      try {
+        const response = await handler(request);
+        return new Response(JSON.stringify({ success: true, ...response }), {
+          headers: {
+            "Content-Type": "application/json",
+            "Access-Control-Allow-Origin": "*",
+          },
+        });
+      } catch (error) {
+        return new Response(JSON.stringify({ success: false, error: error.message }), {
+          status: 500,
+          headers: {
+            "Content-Type": "application/json",
+            "Access-Control-Allow-Origin": "*",
+          },
+        });
+      }
+    }
+  
+    static async parseRequest(request) {
+      const contentType = request.headers.get("content-type") || "";
+      if (contentType.includes("application/json")) {
+        return await request.json();
+      } else if (contentType.includes("application/x-www-form-urlencoded")) {
+        const formData = await request.formData();
+        const data = {};
+        formData.forEach((value, key) => {
+          data[key] = value;
+        });
+        return data;
+      }
+      throw new Error("Unsupported content type");
+    }
+  }
+
